requirements.txt
requests==2.32.3
PyYAML==6.0.2
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
streamlit==1.41.1
pandas==2.2.3
python-dateutil==2.9.0.post0